backend/pyproject.toml
[project]
name = "documaker-backend"
version = "0.1.0"
description = "Backend for Documaker application"
requires-python = ">=3.8"
dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.7",
    # Authentication & Security
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.1.0",
    # HTTP & Templating
    "httpx>=0.25.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    # PDF Generation
    "playwright>=1.40.0",
    # External APIs
    "supabase>=2.0.0",
    # Media handling
    "cloudinary>=1.36.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "*.pyc",
    "__pycache__",
    ".env",
    ".env.*",
    "uploads/",
    "migrations/",
    "tests/"
]
